day20: part2
Struggled with this one a bit trying to come up with a truly general
solution. Should've known better given the lack of examples. With the
given input, there is only one input to the 'rx' module and that input
is a conjunction. Additionally, it turns out that each of the inputs to
that conjunction module form a cycle which has a repeating pattern
starting at an offset button_presses=0. So to find the answer, we just
need to find the number of button presses it takes for each input to the
conjunction to send a 'high' pulse then get the lcm. Definitely not as
general as I would prefer, but I did end up checking the reddit on this
one to confirm that other inputs followed this pattern.